ID system implemented (Joe, ID is now an int, not a long)

+++ TheatreProjectWeb/src/theatreProject/shared/Persistence.java
@@ -16,7 +16,7 @@ public interface Persistence extends RemoteService {
 	User getUser(String email);
 	void saveUser(User user);
 	
-	InventoryObject getInventoryObject(Long ID);
+	InventoryObject getInventoryObject(int ID);
 	void saveObject(InventoryObject object);
 	
 	ArrayList<InventoryObject> search(String parameter);
